+++ kibot/misc.py
@@ -175,6 +175,7 @@ W_TIMEOUT = '(W054) '
 W_MUSTBEINT = '(W055) '
 W_NOOUTPUTS = '(W056) '
 W_NOTASCII = '(W057) '
+W_KIAUTO = '(W058) '
 
 
 class Rect(object):