+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `erc_warnings` pre-flight option to consider ERC warnings as errors.
 
+### Changed
+- Errors and warnings from KiAuto now are printed as errors and warnings.
+
 ### Fixed
 - Problem when using E/DRC filters and the output dir didn't exist.